Move GridFactory into the content app
