+++ content/webapp/components/Body/Body.tsx
@@ -38,9 +38,7 @@ import TagsGroup from '@weco/common/views/components/TagsGroup/TagsGroup';
 import Discussion from '@weco/common/views/components/Discussion/Discussion';
 import WobblyEdgedContainer from '@weco/common/views/components/WobblyEdgedContainer/WobblyEdgedContainer';
 import WobblyEdge from '@weco/common/views/components/WobblyEdge/WobblyEdge';
-import GridFactory, {
-  sectionLevelPageGrid,
-} from '@weco/common/views/components/GridFactory/GridFactory';
+import GridFactory, { sectionLevelPageGrid } from './GridFactory';
 import Card from '@weco/common/views/components/Card/Card';
 import { convertItemToCardProps } from '@weco/common/model/card';
 import { BodyType } from '@weco/common/model/generic-content-fields';
